fix unit tests and added mock vorpal wrapper
